Refactor countdown component
